Add start of the cart system

+++ M32COM - Coursework/Global.asax.cs
@@ -18,7 +18,7 @@ namespace M32COM___Coursework
 
         protected void Session_Start(object sender, EventArgs e)
         {
-            var currentCart = new List<float>();
+            var currentCart = new List<int>();
             //Set no user to be currently logged in
             Session["LoggedIn"] = false;
             Session["CurrentUser"] = null;
